Add files via upload
This is written in such a way that it prints out the sum of the prime factorials for non-prime numbers, where k == 2. In other words, n is the product of two prime numbers.

Note: The prime factorials of 8 are 2, 2, and 2; therefore k == 3. Therefore nothing will be printed for 8.

An example of a solution execution:

% java Prime
The sum of all primes for 4:    4    (2 + 2)
The sum of all primes for 6:    5    (2 + 3)
The sum of all primes for 9:    6    (3 + 3)
The sum of all primes for 10:   7    (2 + 5)
The sum of all primes for 14:   9    (2 + 7)
The sum of all primes for 15:   8    (3 + 5)