Add files via upload
This assignment is to create a word search application using the Scanner class.

My program will prompt the user for an input file that will contain the word search puzzle. The program presents that puzzle to the user and asks which word to search for. The search is able to find all occurrences of a word horizontally, and vertically, both forwards and backward.

If the file entered by the user is not present, you must inform the user and prompt again. If the user enters âexitâ at this prompt, the program should exit with code 0.