Add files via upload
This program prompts the user for an input file that will contain the words to search. The program then presents that puzzle to the user and ask which word to search for. The program searches all occurrences of a word horizontally, and vertically, both forwards and backward. The word can include the following special characters also!!