Add files via upload
The assignment is to create a program similar to cut. I have implemented the arguments -f and -d The command-line argument -f1,3 means you have to print field 1 to 3, and -dxyx means the delimiter between two fields is the string xyx. The code reads from stdin, if the fileName is not specified on command line.

The syntax for calling the program is:

                    java Cut [-fn,k] [-dtext] [-input fileName]
n and k are numbers, text is a string.

cat cut.txt
                    1:2:3:4:5
                    aaa:bb:c
                    % java Cut -f1,2 -d: -input cut.txt
                    1 2
                    aaa bb
                    %